Features/as49 as54

See merge request ashdod-bonded/190709/solocate/ios-driver!2

+++ DMSDriver/DMSDriver/App/ViewControllers/Order/LoadUnloadOrderVC/Cell/LoadUnLoadListCell.swift
@@ -70,7 +70,7 @@ extension LoadUnLoadListCell:UITextFieldDelegate {
     func textFieldDidEndEditing(_ textField: UITextField) {
         if textField == actualQtyTextField, let text = textField.text {
             let validText = (text.isEmpty) ? "0" : text
-            let quantity = Double(validText)
+            let quantity = Int(validText)
             orderDetail?.actualQty = quantity
         }
     }
